Bug floating-point fix

+++ src/main/java/advmain/AdvGamePanel.java
@@ -24,7 +24,7 @@ public class AdvGamePanel extends JPanel implements Runnable{
     public final int worldWidth = tileSize * maxWorldCol;
     public final int worldHeight = tileSize * maxWorldRow;
     // FPS
-    int FPS = 60;
+    double FPS = 60d;
 
     AdvTileManager tileM = new AdvTileManager(this);
     AdvKeyHandler keyH = new AdvKeyHandler();
